Speed up Dymola compilation for identic calls
With Dymola 2025x Refresh 1 the compilation time can be drastically
reduced using this annotation when the same value is read many times to
set parameters.

It allows Dymola to cache the result of the function call and treat the
function as pure when reading values for parameters, as long as no
regular impure function (one without this annotation) is called
in-between.

This is e.g. beneficial for Battery packs with hundreds of cells, where
each cell of an array of cells reads the same parameter values from an
sdf file, like the number of RC elements they should use.